pyproject.toml
[tool.poetry]
name = "api"
version = "0.2.0"
description = "LASUR Web Services"
authors = ["ymarcon <yannick.marcon@epfl.ch>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
typo-modal = {git = "https://github.com/EPFL-LASUR/mobility-toolkit.git", rev = "0.1.2"}
pydantic-settings = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
